Abstract SearchBar related components away from FindChain page
